+++ rotasi.h
@@ -2,4 +2,4 @@
 
 point rotasi(point p1, point p2, double degree);
 
-void rotasiBanyak(point p0, point* p1, int degree, double banyak);
+void rotasiBanyak(point p0, point* p1, double degree, double banyak);
